requirements.txt
httpx[http2]
beautifulsoup4
lxml
python-dotenv
starlette
uvicorn
